move setup and cleanup to a folder and skip that folder
