+++ tests/assets/setup_test.go
@@ -1,7 +1,7 @@
 //go:build e2e
 // +build e2e
 
-package tests
+package assets
 
 import (
 	"context"
