+++ tests/run-arm-smoke-tests.sh
@@ -19,7 +19,7 @@ failed_lookup=()
 counter=0
 
 function run_setup {
-    go test -v -tags e2e setup_test.go
+    go test -v -tags e2e assets/setup_test.go
 }
 
 function run_tests {
@@ -126,7 +126,7 @@ function print_logs {
 }
 
 function run_cleanup {
-    go test -v -tags e2e cleanup_test.go
+    go test -v -tags e2e assets/cleanup_test.go
 }
 
 function print_failed {
